apps/service/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "finnhub-python>=2.4.24",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
]
